Managed card service, collect info from products and add it to card of specific user. Problems with: -integration, -login cant handle email as username, -OrderIntegrationConfig("@Qualifier("ordersChannel")"). Need to create html templates for card and profile manager. Need to create service to add images to products.

+++ Web SIte/src/main/java/springeshop/website/domain/Products.java
@@ -23,8 +23,8 @@ public class Products {
     private Long id;
 
     private String title;
-    private BigDecimal price;
-    private Integer numbers;
+    private Double price;
+    private Long numbers;
 
     @ManyToMany(cascade = CascadeType.ALL)
     @JoinTable(name = "product_categories",
@@ -35,7 +35,7 @@ public class Products {
     public Products() {
     }
 
-    public Products(String title, BigDecimal price, Integer numbers) {
+    public Products(String title, Double price, Long numbers) {
         this.title = title;
         this.price = price;
         this.numbers = numbers;
